requirements.txt
# Core Dependencies
langchain-core>=0.1.27
langchain-openai>=0.0.8
langchain-anthropic>=0.1.4
langchain-google-genai>=0.0.11
google-generativeai>=0.3.2
anthropic>=0.18.1
openai>=1.12.0
langchain-community>=0.0.10

# AI and ML
sentence-transformers>=2.5.1
chromadb>=0.4.24
torch>=2.0.0
transformers>=4.30.0

# Desktop Automation
opencv-python>=4.9.0.80
pyautogui>=0.9.54
keyboard>=0.13.5
pillow>=10.2.0

# Database and Storage
aiosqlite>=0.20.0
sqlalchemy>=2.0.0

# System Monitoring
psutil>=5.9.0

# Security and Configuration
python-dotenv>=1.0.0
cryptography>=42.0.5

# Async Support
aiohttp>=3.9.3
asyncio>=3.4.3

# Testing
pytest>=8.0.2
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0

# ユーティリティ
pydantic>=2.6.3
numpy>=1.24.0
pandas>=2.1.0

# 開発ツール
black>=22.0.0
flake8>=4.0.0

# システムモニタリング
GPUtil>=1.4.0

# ロギング
python-json-logger==2.0.7

# 分析関連
seaborn>=0.12.0
pyyaml>=6.0.0

# システムモニタリング
wmi>=1.5.1
pynvml>=11.5.0

# マルチメディア処理
dash>=2.14.0
fastapi>=0.109.0
optuna>=3.5.0
prophet>=1.1.5
plotly>=5.18.0

# ユーティリティ
jinja2>=3.1.0
aiosmtplib>=3.0.0
aiofiles>=23.2.0

# Vision Processing
pytesseract>=0.3.10
mediapipe>=0.10.9
dlib>=19.24.2

# Added from the code block
PyQt6>=6.4.0
PyQt6-WebEngine>=6.4.0
python-dotenv>=1.0.0
psutil>=5.9.0
GPUtil>=1.4.0
WMI>=1.5.1
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.6
opencv-python>=4.9.0
keyboard>=0.13.5
pyautogui>=0.9.54
Pillow>=10.0.0
pandas>=2.0.0 